requirements.txt
# Core
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Voice
edge-tts>=6.1.0
faster-whisper>=1.0.0
pydub>=0.25.0

# Async
anyio>=4.0.0
asyncio-throttle>=1.0.0

# Utilities
aiofiles>=23.0.0
httpx>=0.26.0
ddgs>=9.0.0
